+++ src/GoogleDriveDocsBundle.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Borsche\GoogleDriveDocsBundle;
 
 use Borsche\GoogleDriveDocsBundle\DependencyInjection\Compiler\ValidateCachePoolPass;
+use Borsche\GoogleDriveDocsBundle\DependencyInjection\Compiler\WarnDefaultViewerContextPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -15,5 +16,6 @@ class GoogleDriveDocsBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new ValidateCachePoolPass());
+        $container->addCompilerPass(new WarnDefaultViewerContextPass());
     }
 }